requirements.txt
Flask
Flask-Cors
Flask-SocketIO
gunicorn
python-dotenv
google-generativeai
openai
supabase
beautifulsoup4
requests
selenium
webdriver_manager


